+++ src/main/java/ok/dht/test/kurdyukov/db/base/BaseEntry.java
