+++ src/main/java/ok/dht/test/kurdyukov/db/base/Dao.java
@@ -84,5 +84,4 @@ public interface Dao<D, E extends Entry<D>> extends Closeable {
     default void close() throws IOException {
         flush();
     }
-
 }
