+++ src/main/java/ok/dht/test/kurdyukov/db/base/Entry.java
