+++ src/main/java/ok/dht/test/kurdyukov/db/dao/EntryKeyComparator.java
@@ -1,16 +1,15 @@
-package ok.dht.test.kurdyukov.db.storage;
+package ok.dht.test.kurdyukov.db.dao;
 
 import jdk.incubator.foreign.MemorySegment;
 import ok.dht.test.kurdyukov.db.base.Entry;
 
 import java.util.Comparator;
 
-public class EntryKeyComparator implements Comparator<Entry<MemorySegment>> {
+public final class EntryKeyComparator implements Comparator<Entry<MemorySegment>> {
 
     public static final Comparator<Entry<MemorySegment>> INSTANCE = new EntryKeyComparator();
 
-    private EntryKeyComparator() {
-    }
+    private EntryKeyComparator() {}
 
     @Override
     public int compare(Entry<MemorySegment> o1, Entry<MemorySegment> o2) {
