+++ src/main/java/ok/dht/test/kurdyukov/db/exception/StorageClosedException.java
@@ -5,5 +5,4 @@ public class StorageClosedException extends RuntimeException {
     public StorageClosedException(Throwable causedBy) {
         super(causedBy);
     }
-
 }
